Add basic coinflip command
Implement a new command 'coinflip' that allows users to flip a coin and get the result of heads or tails. The command is added to the 'interactionCreate' event handler in 'app.js'. The 'coinflip.js' file is also added to the 'commands' directory. The 'coinFlip' function is implemented in 'utils/utils.js' to generate a random result of heads or tails. This change enhances the functionality of the application.